requirements.txt
flask
flask-cors
tensorflow
opencv-python
Pillow
huggingface_hub
requests
google-generativeai
